CV 2.0 completado
