+++ src/router/routes.js
@@ -4,7 +4,12 @@ const routes = [
     path: '/',
     component: () => import('layouts/MainLayout.vue'),
     children: [
-      { path: '', component: () => import('pages/IndexPage.vue') }
+      { path: '', component: () => import('pages/InicioPage.vue') },
+      { path: '/skills', component: () => import('pages/HabilidadesPage.vue') },
+      { path: '/about', component: () => import('pages/SobrePage.vue') },
+      { path: '/projects', component: () => import('pages/ProyectosPage.vue') },
+      { path: '/contact', component: () => import('pages/ContactosPage.vue') }
+
     ]
   },
